% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\docType{data}
\name{centenarians}
\alias{centenarians}
\title{Data Documentation: centenarians dataset}
\format{
A data frame with people who aged 110 years above.
\itemize{
  \item rank
  \item name
  \item birth_date
  \item death_date
  \item age
  \item place_of_death_or_residence
  \item gender
  \item still_alive
}
}
\description{
This dataset contains information about breeds.
}
\keyword{datasets}
